+++ components/sql/src/main/scala/pl/touk/nussknacker/sql/service/DatabaseEnricherInvoker.scala
@@ -30,16 +30,15 @@ class DatabaseEnricherInvoker(query: String,
   override def invokeService(params: Map[String, Any])
                             (implicit ec: ExecutionContext, collector: ServiceInvocationCollector, contextId: ContextId, componentUseCase: ComponentUseCase): Future[queryExecutor.QueryResult] =
     getTimeMeasurement().measuring {
-      Future.successful {
-        queryDatabase(queryArgumentsExtractor(argsCount, params))
-      }
+      queryDatabase(queryArgumentsExtractor(argsCount, params))
     }
 
-  protected def queryDatabase(queryArguments: QueryArguments): queryExecutor.QueryResult =
+  protected def queryDatabase(queryArguments: QueryArguments)(implicit ec: ExecutionContext): Future[queryExecutor.QueryResult] = Future {
     withConnection(query) { statement =>
       setQueryArguments(statement, queryArguments)
       queryExecutor.execute(statement)
     }
+  }
 
   protected def setQueryArguments(statement: PreparedStatement, queryArguments: QueryArguments): Unit =
     queryArguments.value.foreach { arg =>