+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkProcessTestRunner.scala
@@ -7,13 +7,13 @@ import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.testmode.TestProcess.TestResults
 import pl.touk.nussknacker.engine.util.StaticMethodRunner
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
 class FlinkProcessTestRunner(modelData: ModelData) extends StaticMethodRunner(modelData.modelClassLoader.classLoader,
   "pl.touk.nussknacker.engine.process.runner.FlinkTestMain", "run") {
 
-  def test[T](canonicalProcess: CanonicalProcess, scenarioTestData: ScenarioTestData, variableEncoder: Any => T): Future[TestResults[T]] = {
-    Future.successful(tryToInvoke(modelData, canonicalProcess, scenarioTestData, new Configuration(), variableEncoder).asInstanceOf[TestResults[T]])
+  def test[T](canonicalProcess: CanonicalProcess, scenarioTestData: ScenarioTestData, variableEncoder: Any => T)(implicit ec: ExecutionContext): Future[TestResults[T]] = Future {
+    tryToInvoke(modelData, canonicalProcess, scenarioTestData, new Configuration(), variableEncoder).asInstanceOf[TestResults[T]]
   }
 
 }